+++ Src/FluentAssertions/Equivalency/SelfReferenceEquivalencyAssertionOptions.cs
@@ -317,6 +317,18 @@ namespace FluentAssertions.Equivalency
         }
 
         /// <summary>
+        /// Includes the specified member in the equality check.
+        /// </summary>
+        /// <remarks>
+        /// This overrides the default behavior of including all declared members.
+        /// </remarks>
+        public TSelf Including(Expression<Func<IMemberInfo, bool>> predicate)
+        {
+            AddSelectionRule(new IncludeMemberByPredicateSelectionRule(predicate));
+            return (TSelf)this;
+        }
+
+        /// <summary>
         /// Tries to match the members of the subject with equally named members on the expectation. Ignores those
         /// members that don't exist on the expectation and previously registered matching rules.
         /// </summary>
@@ -698,6 +710,7 @@ namespace FluentAssertions.Equivalency
             /// <typeparamref name="TMemberType" />
             /// </summary>
             public TSelf WhenTypeIs<TMemberType>()
+                where TMemberType : TMember
             {
                 When(info => info.RuntimeType.IsSameOrInherits(typeof(TMemberType)));
                 return options;