+++ Src/FluentAssertions/Formatting/Formatter.cs
@@ -47,6 +47,7 @@ namespace FluentAssertions.Formatting
             new ExceptionValueFormatter(),
             new MultidimensionalArrayFormatter(),
             new EnumerableValueFormatter(),
+            new EnumValueFormatter(),
             new DefaultValueFormatter()
         };